lumos_learning/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-allauth==0.57.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
Pillow==10.1.0
python-decouple==3.8
paypalrestsdk==1.13.3
requests==2.31.0
django-csp==3.7
django-ratelimit==4.1.0
whitenoise==6.6.0
boto3==1.34.0
django-storages==1.14.2
coverage==7.3.2